tests moving context manager to avoid circular imports
